% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvm.R
\name{coef.tsvm}
\alias{coef.tsvm}
\title{Extract Twin-SVM Coefficients}
\usage{
\method{coef}{tsvm}(object, ...)
}
\arguments{
\item{object}{A fitted \code{tsvm} object.}

\item{...}{Unused.}
}
\value{
A list with the two plane coefficients.
}
\description{
Extract Twin-SVM Coefficients
}
\examples{
dat <- gen_moons(30)
fit <- tsvm(dat$x, dat$y)
coef(fit)
}
