% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{gen_circles}
\alias{gen_circles}
\title{Generate Concentric Circles Data}
\usage{
gen_circles(n, noise = 0.05)
}
\arguments{
\item{n}{Number of observations.}

\item{noise}{Standard deviation of Gaussian noise.}
}
\value{
A list with numeric matrix \code{x} and factor \code{y}.
}
\description{
Generate Concentric Circles Data
}
\examples{
dat <- gen_circles(20, noise = 0.05)
str(dat)
}
