% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{gen_moons}
\alias{gen_moons}
\title{Generate Two-Moons Data}
\usage{
gen_moons(n, noise = 0.2)
}
\arguments{
\item{n}{Number of observations.}

\item{noise}{Standard deviation of Gaussian noise.}
}
\value{
A list with numeric matrix \code{x} and factor \code{y}.
}
\description{
Generate Two-Moons Data
}
\examples{
dat <- gen_moons(20, noise = 0.1)
str(dat)
}
