% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{plot.cv_tsvm}
\alias{plot.cv_tsvm}
\title{Plot Twin-SVM Cross-Validation Results}
\usage{
\method{plot}{cv_tsvm}(x, ...)
}
\arguments{
\item{x}{A \code{cv_tsvm} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Twin-SVM Cross-Validation Results
}
\examples{
set.seed(11)
dat <- gen_moons(40, noise = 0.1)
cv <- cv_tsvm(dat$x, dat$y, c1_grid = c(0.1, 1), c2_grid = c(0.1, 1), k = 3)
plot(cv)
}
