% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.svms}
\alias{plot.svms}
\title{Plot a Standard SVM Decision Boundary}
\usage{
\method{plot}{svms}(x, ...)
}
\arguments{
\item{x}{A fitted \code{svms} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a Standard SVM Decision Boundary
}
\examples{
dat <- gen_moons(40)
fit <- svms(dat$x, dat$y, kernel = "linear")
plot(fit)
}
