% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tsvm}
\alias{plot.tsvm}
\title{Plot a Twin-SVM Decision Boundary}
\usage{
\method{plot}{tsvm}(x, ...)
}
\arguments{
\item{x}{A fitted \code{tsvm} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a Twin-SVM Decision Boundary
}
\examples{
dat <- gen_moons(40)
fit <- tsvm(dat$x, dat$y)
plot(fit)
}
