% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovo.R
\name{print.svms_multiclass}
\alias{print.svms_multiclass}
\title{Print a Multiclass Standard SVM}
\usage{
\method{print}{svms_multiclass}(x, ...)
}
\arguments{
\item{x}{A fitted \code{svms_multiclass} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a Multiclass Standard SVM
}
\examples{
set.seed(46)
x <- rbind(
  matrix(rnorm(8, -2, 0.2), ncol = 2),
  matrix(rnorm(8, 0, 0.2), ncol = 2),
  matrix(rnorm(8, 2, 0.2), ncol = 2)
)
y <- factor(rep(c("a", "b", "c"), each = 4))
print(svms(x, y, kernel = "linear", max_passes = 2, max_iter = 100))
}
\seealso{
Other multiclass:
\code{\link[=confusion]{confusion()}},
\code{\link[=predict.svms_multiclass]{predict.svms_multiclass()}},
\code{\link[=predict.tsvm_multiclass]{predict.tsvm_multiclass()}},
\code{\link[=print.tsvm_multiclass]{print.tsvm_multiclass()}}
}
\concept{multiclass}
