% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovo.R
\name{print.tsvm_multiclass}
\alias{print.tsvm_multiclass}
\title{Print a Multiclass Twin SVM}
\usage{
\method{print}{tsvm_multiclass}(x, ...)
}
\arguments{
\item{x}{A fitted \code{tsvm_multiclass} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a Multiclass Twin SVM
}
\examples{
set.seed(41)
x <- rbind(
  matrix(rnorm(20, -2, 0.2), ncol = 2),
  matrix(rnorm(20, 0, 0.2), ncol = 2),
  matrix(rnorm(20, 2, 0.2), ncol = 2)
)
y <- factor(rep(c("a", "b", "c"), each = 10))
print(tsvm(x, y, kernel = "linear"))
}
\seealso{
Other multiclass:
\code{\link[=confusion]{confusion()}},
\code{\link[=predict.svms_multiclass]{predict.svms_multiclass()}},
\code{\link[=predict.tsvm_multiclass]{predict.tsvm_multiclass()}},
\code{\link[=print.svms_multiclass]{print.svms_multiclass()}}
}
\concept{multiclass}
