% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm-package.R
\docType{package}
\name{twinsvm-package}
\alias{twinsvm}
\alias{twinsvm-package}
\title{twinsvm: Twin Support Vector Machines}
\description{
Twin support vector machine classifiers with a standard C-SVC baseline.
}
\references{
Jayadeva, Khemchandani, R., and Chandra, S. (2007). Twin support vector
machines for pattern classification. \emph{IEEE Transactions on Pattern Analysis
and Machine Intelligence}, 29(5), 905-910.

Kumar, M. A. and Gopal, M. (2009). Least squares twin support vector
machines for pattern classification. \emph{Expert Systems with Applications},
36(4), 7535-7543.
}
\author{
\strong{Maintainer}: Shamika Tissera \email{nimeshshamika@gmail.com}

Authors:
\itemize{
  \item Shamika Tissera \email{nimeshshamika@gmail.com}
}

}
