% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twomoons}
\alias{twomoons}
\title{Example Two-Moons Data}
\format{
A data frame with 120 rows and 3 variables:
\describe{
\item{x1}{First coordinate.}
\item{x2}{Second coordinate.}
\item{y}{Two-class factor with levels \code{B} and \code{A}.}
}
}
\usage{
data(twomoons)
}
\description{
A small two-class two-moons data set for examples and tests.
}
\examples{
data(twomoons)
str(twomoons)
}
\keyword{datasets}
